requirements.txt
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
